requirements.txt
accelerate==1.9.0
einops==0.8.1
matplotlib==3.7.0
networkx==3.2.1
numpy==1.26.4
pandas==2.2.2
pyg-lib==0.4.0+pt20cu117
torch==2.0.0
torch-cluster==1.6.3+pt20cu117
torch-geometric==2.6.1
torch-scatter==2.1.2+pt20cu117
torch-sparse==0.6.18+pt20cu117
torch-spline-conv==1.2.2+pt20cu117
pytorch-lightning==2.0.0
PyYAML==6.0.2
scikit-learn==1.6.1
scipy==1.11.4
seaborn==0.13.2
setproctitle==1.2.3
tensorboard==2.9.0
torchdiffeq==0.2.5
torchdyn==1.0.6
torchsde==0.2.6
torchcfm==1.0.7
torchmetrics==1.8.0
torchsummary==1.5.1
tqdm==4.64.0
wandb==0.23.1